add custom auth. Fixes

+++ src/views/message_router.py
@@ -1,6 +1,7 @@
 import json
 import uuid
 
+from src.app import validate
 from src.controllers.message_controller import MessageController
 
 import boto3
@@ -11,15 +12,14 @@ from src.helpers import DEBUG
 router = Router()
 
 
-# @router.post("/")
-# def test():
-#     random_id = uuid.uuid4().__str__()
-#     return random_id
-
-
 @router.post("/messages")
 def post_message():
     event = router.current_event
+
+    auth = validate(event.get("headers", {}).get("Authorization"))
+    if not auth:
+        return "Unauthorized.Sorry"
+
     DEBUG(f"POST REQUEST: {event}")
 
     body = json.loads(event.body)
@@ -32,6 +32,12 @@ def post_message():
 
 @router.get("/messages")
 def get_messages():
+    event = router.current_event
+
+    auth = validate(event.get("headers", {}).get("Authorization"))
+    if not auth:
+        return "Unauthorized.Sorry"
+
     controller = MessageController()
     result = controller.get_messages_from_db()
     return result
@@ -39,6 +45,12 @@ def get_messages():
 
 @router.get("/messages/<message_id>")
 def get_message_by_id(message_id):
+    event = router.current_event
+
+    auth = validate(event.get("headers", {}).get("Authorization"))
+    if not auth:
+        return "Unauthorized.Sorry"
+
     controller = MessageController()
     result = controller.get_message_from_db_by_id(message_id=message_id)
     return result
@@ -46,6 +58,12 @@ def get_message_by_id(message_id):
 
 @router.delete("/messages")
 def delete_messages():
+    event = router.current_event
+
+    auth = validate(event.get("headers", {}).get("Authorization"))
+    if not auth:
+        return "Unauthorized.Sorry"
+
     controller = MessageController()
     result = controller.delete_messages_from_db()
     return result
@@ -53,6 +71,12 @@ def delete_messages():
 
 @router.delete("/messages/<message_id>")
 def delete_message_by_id(message_id):
+    event = router.current_event
+
+    auth = validate(event.get("headers", {}).get("Authorization"))
+    if not auth:
+        return "Unauthorized.Sorry"
+
     controller = MessageController()
     result = controller.delete_message_from_db_by_id(message_id=message_id)
     return result
